refactor(path): 임시 클래스들을 제거했다.
