+++ src/main/java/nextstep/subway/path/domain/DistanceSurchargePolicy.java
@@ -16,7 +16,7 @@ public class DistanceSurchargePolicy extends SurchargePolicy {
   }
 
   @Override
-  protected long getSurchargeAmount(Path2 path) {
+  protected long getSurchargeAmount(Path path) {
     long distance = Math.min(path.getTotalDistance(), upperBound);
     return (long) Math.ceil((double) (distance - lowerBound) / distanceUnit) * farePerUnitDistance;
   }
