+++ src/main/java/nextstep/subway/path/ui/PathController.java
@@ -6,10 +6,10 @@ import nextstep.auth.ui.AuthenticationPrincipal;
 import nextstep.member.application.MemberService;
 import nextstep.member.domain.Member;
 import nextstep.subway.path.application.FareCalculator;
-import nextstep.subway.path.application.PathService2;
+import nextstep.subway.path.application.PathService;
 import nextstep.subway.path.application.dto.PathRequest;
 import nextstep.subway.path.application.dto.PathResponse;
-import nextstep.subway.path.domain.Path2;
+import nextstep.subway.path.domain.Path;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -18,7 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 public class PathController {
-  private final PathService2 pathService;
+  private final PathService pathService;
   private final FareCalculator fareCalculator;
   private final MemberService memberService;
 
@@ -26,7 +26,7 @@ public class PathController {
   public ResponseEntity<PathResponse> findPath(
       @ModelAttribute PathRequest request, @AuthenticationPrincipal LoginMember loginMember) {
     Member member = memberService.findMemberByEmail(loginMember.getEmail());
-    Path2 path = pathService.findPath(request);
+    Path path = pathService.findPath(request);
     long fare = fareCalculator.calculateFare(path, member);
     return ResponseEntity.ok(PathResponse.of(path, fare));
   }
