+++ src/test/java/nextstep/subway/unit/path/application/GraphServiceTest.java
@@ -1,20 +1,21 @@
 package nextstep.subway.unit.path.application;
 
-import static nextstep.Fixtures.강남_역삼_구간;
-import static nextstep.Fixtures.이호선;
+import static nextstep.Fixtures.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 import java.util.List;
 import nextstep.subway.line.application.LineService;
 import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.domain.LineSection;
 import nextstep.subway.path.application.GraphService;
-import nextstep.subway.path.domain.*;
-import nextstep.subway.station.domain.Station;
+import nextstep.subway.path.domain.PathType;
+import nextstep.subway.path.domain.SubwayGraph;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -26,39 +27,22 @@ class GraphServiceTest {
   @Mock LineService lineService;
   @InjectMocks GraphService graphService;
 
-  @DisplayName("노선 정보 바탕으로 거리 그래프를 불러온다.")
-  @Test
-  void loadGraphDistance() {
+  @DisplayName("노선 정보 바탕으로 그래프를 불러온다.")
+  @ParameterizedTest
+  @EnumSource(PathType.class)
+  void loadGraph(PathType type) {
     Line 이호선 = 이호선();
-    LineSection 강남_역삼_구간 = 강남_역삼_구간();
-    Station 강남역 = 강남_역삼_구간.getUpStation();
-    Station 역삼역 = 강남_역삼_구간.getDownStation();
-    given(lineService.findAllLines()).willReturn(List.of(이호선));
+    Line 신분당선 = 신분당선();
+    given(lineService.findAllLines()).willReturn(List.of(이호선, 신분당선));
 
-    SubwayGraph graph = graphService.loadGraph(PathType.DISTANCE);
+    SubwayGraph graph = graphService.loadGraph(type);
 
-    SubwayGraph expectedGraph = new SubwayGraph(PathType.DISTANCE);
-    expectedGraph.addStation(강남역);
-    expectedGraph.addStation(역삼역);
-    expectedGraph.addLineSection(강남_역삼_구간);
-    assertThat(graph.isSame(expectedGraph)).isTrue();
-  }
+    verify(lineService, times(1)).findAllLines();
 
-  @DisplayName("노선 정보 바탕으로 소요시간 그래프를 불러온다.")
-  @Test
-  void loadGraphDuration() {
-    Line 이호선 = 이호선();
-    LineSection 강남_역삼_구간 = 강남_역삼_구간();
-    Station 강남역 = 강남_역삼_구간.getUpStation();
-    Station 역삼역 = 강남_역삼_구간.getDownStation();
-    given(lineService.findAllLines()).willReturn(List.of(이호선));
+    SubwayGraph expectedGraph = new SubwayGraph(type);
+    expectedGraph.addLine(이호선);
+    expectedGraph.addLine(신분당선);
 
-    SubwayGraph graph = graphService.loadGraph(PathType.DURATION);
-
-    SubwayGraph expectedGraph = new SubwayGraph(PathType.DURATION);
-    expectedGraph.addStation(강남역);
-    expectedGraph.addStation(역삼역);
-    expectedGraph.addLineSection(강남_역삼_구간);
     assertThat(graph.isSame(expectedGraph)).isTrue();
   }
 }
