+++ src/test/java/nextstep/subway/unit/path/application/dto/PathResponseTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.Arrays;
 import java.util.List;
 import nextstep.subway.path.application.dto.PathResponse;
-import nextstep.subway.path.domain.Path2;
+import nextstep.subway.path.domain.Path;
 import nextstep.subway.station.application.dto.StationResponse;
 import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.DisplayName;
@@ -21,7 +21,7 @@ class PathResponseTest {
     Station 교대역 = 교대역();
     Station 남부터미널역 = 남부터미널역();
     Station 양재역 = 양재역();
-    Path2 path = Path2.of(Arrays.asList(교대역, 남부터미널역, 양재역), List.of(삼호선(), 이호선()), 5, 10);
+    Path path = Path.of(Arrays.asList(교대역, 남부터미널역, 양재역), List.of(삼호선(), 이호선()), 5, 10);
     long fare = 1250;
 
     PathResponse response = PathResponse.of(path, fare);
